cleaned up the main file. create new classes to reduce the code lines in the main file
